refactror: simplify permission package by having the logic only inside the hook
