+++ packages/permissions/src/components/branch-on-permissions/branch-on-permissions.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { ApplicationContext } from '@commercetools-frontend/application-shell-connectors';
 import getDisplayName from '../../utils/get-display-name';
 import Authorized from '../authorized';
 
@@ -42,31 +41,23 @@ const branchOnPermissions = <OwnProps extends {}>(
   Component: React.ComponentType<OwnProps>
 ): React.ComponentType<OwnProps> => {
   const WrappedComponent = (props: OwnProps) => (
-    <ApplicationContext
-      render={applicationContext => (
-        <Authorized
-          shouldMatchSomePermissions={options.shouldMatchSomePermissions}
-          actualPermissions={applicationContext.permissions}
-          actualActionRights={applicationContext.actionRights}
-          actualDataFences={applicationContext.dataFences}
-          demandedPermissions={demandedPermissions}
-          demandedActionRights={options.actionRights}
-          demandedDataFences={options.dataFences}
-          selectDataFenceData={
-            options.getSelectDataFenceData &&
-            options.getSelectDataFenceData(props)
-          }
-          render={isAuthorized => {
-            if (isAuthorized) {
-              return <Component {...props} />;
-            }
-            if (FallbackComponent) {
-              return <FallbackComponent />;
-            }
-            return <React.Fragment></React.Fragment>;
-          }}
-        />
-      )}
+    <Authorized
+      shouldMatchSomePermissions={options.shouldMatchSomePermissions}
+      demandedPermissions={demandedPermissions}
+      demandedActionRights={options.actionRights}
+      demandedDataFences={options.dataFences}
+      selectDataFenceData={
+        options.getSelectDataFenceData && options.getSelectDataFenceData(props)
+      }
+      render={isAuthorized => {
+        if (isAuthorized) {
+          return <Component {...props} />;
+        }
+        if (FallbackComponent) {
+          return <FallbackComponent />;
+        }
+        return <React.Fragment></React.Fragment>;
+      }}
     />
   );
   WrappedComponent.displayName = `branchOnPermissions(${getDisplayName<
