+++ packages/permissions/src/components/restricted-by-permissions/restricted-by-permissions.spec.tsx
@@ -17,7 +17,6 @@ type TAllAppliedMenuVisibility = {
   value: boolean;
 };
 type TAllAppliedDataFence = {
-  __typename: 'StoreDataFence';
   value: string;
   group: string;
   name: string;
