+++ packages/permissions/src/components/restricted-by-permissions/restricted-by-permissions.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import invariant from 'tiny-invariant';
 import isNil from 'lodash/isNil';
-import { ApplicationContext } from '@commercetools-frontend/application-shell-connectors';
 import Authorized from '../authorized';
 
 const getHasChildren = (children: React.ReactNode) =>
@@ -51,40 +50,33 @@ const RestrictedByPermissions = (props: Props) => {
     '@commercetools-frontend/permissions/RestrictedByPermissions: You provided both `children` and `unauthorizedComponent`. Please provide only one of them.'
   );
   return (
-    <ApplicationContext
-      render={applicationContext => (
-        <Authorized
-          shouldMatchSomePermissions={props.shouldMatchSomePermissions}
-          actualPermissions={applicationContext.permissions}
-          actualActionRights={applicationContext.actionRights}
-          actualDataFences={applicationContext.dataFences}
-          demandedPermissions={props.permissions}
-          demandedActionRights={props.actionRights}
-          demandedDataFences={props.dataFences}
-          selectDataFenceData={props.selectDataFenceData}
-          render={(isAuthorized: boolean) => {
-            if (typeof props.children === 'function')
-              return props.children({
-                isAuthorized,
-              });
-            if (typeof props.render === 'function')
-              return props.render({
-                isAuthorized,
-              });
+    <Authorized
+      shouldMatchSomePermissions={props.shouldMatchSomePermissions}
+      demandedPermissions={props.permissions}
+      demandedActionRights={props.actionRights}
+      demandedDataFences={props.dataFences}
+      selectDataFenceData={props.selectDataFenceData}
+      render={(isAuthorized: boolean) => {
+        if (typeof props.children === 'function')
+          return props.children({
+            isAuthorized,
+          });
+        if (typeof props.render === 'function')
+          return props.render({
+            isAuthorized,
+          });
 
-            if (isAuthorized) {
-              if (props.children && getHasChildren(props.children))
-                return React.Children.only(props.children);
-            } else if (!isAuthorized) {
-              if (props.unauthorizedComponent) {
-                return React.createElement(props.unauthorizedComponent);
-              }
-            }
+        if (isAuthorized) {
+          if (props.children && getHasChildren(props.children))
+            return React.Children.only(props.children);
+        } else if (!isAuthorized) {
+          if (props.unauthorizedComponent) {
+            return React.createElement(props.unauthorizedComponent);
+          }
+        }
 
-            return null;
-          }}
-        />
-      )}
+        return null;
+      }}
     />
   );
 };
